update deploy scripts
