+++ deploy/016_deploy_Yeet24ClaimModule.ts
@@ -1,7 +1,7 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { getNetworkConfig } from "../test/utils";
+// import { getNetworkConfig } from "../test/utils";
 
 const deployFn: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
   const { getChainId, deployments, ethers, network } = hre;
@@ -9,22 +9,30 @@ const deployFn: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
 
   console.log("deployer", deployer);
 
-  const networkConfig = getNetworkConfig();
+  // const networkConfig = getNetworkConfig();
 
   console.log("\nDeploying Yeet24ClaimModule on network:", network.name);
 
-  const initialOwner = "0xCED608Aa29bB92185D9b6340Adcbfa263DAe075b";
-  const hos = "0xde65e8b424438b361d8f4a8896f92956510b08dc";
-  console.log("using hos at this address", hos);
+  const maxReward = "300000000000000000"; // TODO: parametrize
+  const rewardPercent = "15"; // TODO: parametrize
+  let initialOwner = "0xCED608Aa29bB92185D9b6340Adcbfa263DAe075b"; // TODO: parametrize
+  let yeet24HOSAddress = "0xde65e8b424438b361d8f4a8896f92956510b08dc"; // TODO: parametrize
+  if (network.name === "hardhat") {
+    initialOwner = deployer;
+    yeet24HOSAddress = (await deployments.get("Yeet24HOS")).address;
+  }
+
   // shaman template id keccak256(abi.encode("Yeet24ShamanModule"))
   const encodedData = ethers.utils.defaultAbiCoder.encode(["string"], ["Yeet24ShamanModule"]);
   const shamanTemplateId = ethers.utils.keccak256(encodedData);
-  const maxReward = "3000000000000000000";
-  const rewardPercent = "15";
+  
+  const params = [initialOwner, yeet24HOSAddress, shamanTemplateId, maxReward, rewardPercent];
+
+  console.log("Yeet24ShamanModule params", params);
 
   const initParams = ethers.utils.defaultAbiCoder.encode(
     ["address", "address", "bytes32", "uint256", "uint256"],
-    [initialOwner, hos, shamanTemplateId, maxReward, rewardPercent],
+    params,
   );
 
   const yeet24ClaimModuleDeployed = await deployments.deploy("Yeet24ClaimModule", {
@@ -34,7 +42,7 @@ const deployFn: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
     proxy: {
       proxyContract: "UUPS",
       execute: {
-        methodName: "__Yeet24ClaimModule_init",
+        methodName: "initialize",
         args: [initParams],
       },
     },
